+++ mcdull-support/mcdull-uac/mcdull-uac-provider/src/main/java/io/gitee/dqcer/mcdull/uac/provider/model/dto/DictValueUpdateDTO.java
@@ -19,5 +19,5 @@ public class DictValueUpdateDTO extends DictValueAddDTO {
 
     @Schema(description = "valueId")
     @NotNull(message = "valueId不能为空")
-    private Long dictValueId;
+    private Integer dictValueId;
 }
