+++ mcdull-support/mcdull-uac/mcdull-uac-provider/src/main/java/io/gitee/dqcer/mcdull/uac/provider/model/dto/EnterpriseAddDTO.java
@@ -85,7 +85,7 @@ public class EnterpriseAddDTO {
     private Boolean disabledFlag;
 
     @Schema(description = "创建人", hidden = true)
-    private Long createUserId;
+    private Integer createUserId;
 
     @Schema(description = "创建人", hidden = true)
     private String createUserName;
