+++ mcdull-support/mcdull-uac/mcdull-uac-provider/src/main/java/io/gitee/dqcer/mcdull/uac/provider/model/dto/UserAddDTO.java
@@ -36,7 +36,7 @@ public class UserAddDTO implements DTO {
 
     @Schema(description = "部门id")
     @NotNull(message = "部门id不能为空")
-    private Long departmentId;
+    private Integer departmentId;
 
     @Schema(description = "是否启用")
     @NotNull(message = "是否被禁用不能为空")
@@ -47,5 +47,5 @@ public class UserAddDTO implements DTO {
     private String phone;
 
     @Schema(description = "角色列表")
-    private List<Long> roleIdList;
+    private List<Integer> roleIdList;
 }
