+++ mcdull-support/mcdull-uac/mcdull-uac-provider/src/main/java/io/gitee/dqcer/mcdull/uac/provider/model/entity/ChangeLogEntity.java
@@ -17,7 +17,7 @@ import java.util.Date;
 @Getter
 @Setter
 @TableName("sys_change_log")
-public class ChangeLogEntity extends RelEntity<Long> {
+public class ChangeLogEntity extends RelEntity<Integer> {
 
     private static final long serialVersionUID = 1L;
