+++ mcdull-support/mcdull-uac/mcdull-uac-provider/src/main/java/io/gitee/dqcer/mcdull/uac/provider/model/entity/OaEnterpriseEntity.java
@@ -17,7 +17,7 @@ import java.time.LocalDate;
 @Getter
 @Setter
 @TableName("sys_oa_enterprise")
-public class OaEnterpriseEntity extends BaseEntity<Long> {
+public class OaEnterpriseEntity extends BaseEntity<Integer> {
 
     private static final long serialVersionUID = 1L;
